# Small optimization for integers Display implementation