@rust-log-analyzer 2024-07-25
The job **`mingw-check`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/10100551995/job/27932166884) [(plain)](https://github.com/rust-lang/rust/commit/1a01d26e64ac2997f273d7a26f35a95bc16aff71/checks/27932166884/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain
#16 2.812 Building wheels for collected packages: reuse
#16 2.813   Building wheel for reuse (pyproject.toml): started
#16 3.054   Building wheel for reuse (pyproject.toml): finished with status 'done'
#16 3.055   Created wheel for reuse: filename=reuse-4.0.3-cp310-cp310-manylinux_2_35_x86_64.whl size=132715 sha256=dfa09868353292d98f811d3efdb0d54d07389e808efc71d68e3b93c514bf8bec
#16 3.056   Stored in directory: /tmp/pip-ephem-wheel-cache-_9gu5d12/wheels/3d/8d/0a/e0fc6aba4494b28a967ab5eaf951c121d9c677958714e34532
#16 3.058 Installing collected packages: boolean-py, binaryornot, tomlkit, reuse, python-debian, markupsafe, license-expression, jinja2, chardet, attrs
#16 3.455 Successfully installed attrs-23.2.0 binaryornot-0.4.4 boolean-py-4.0 chardet-5.2.0 jinja2-3.1.4 license-expression-30.3.0 markupsafe-2.1.5 python-debian-0.1.49 reuse-4.0.3 tomlkit-0.13.0
#16 3.455 WARNING: Running pip as the 'root' user can result in broken permissions and conflicting behaviour with the system package manager. It is recommended to use a virtual environment instead: https://pip.pypa.io/warnings/venv
#16 DONE 3.5s
---
    |
213 | macro_rules! impl_Display {
    | ------------------------- when calling this macro
...
572 |         i64 => 20 u64 => named fmt_i64,
    |                   ^^^ no rules expected this token in macro call
note: while trying to match `=>`
   --> library/core/src/fmt/num.rs:214:73
    |
    |
214 |     ($($t:ident => $size:literal $(as $positive:ident in $other:ident)? => named $name:ident,)* ; as $u:ident via $conv_fn:ident named $g...

error[E0277]: `i64` doesn't implement `fmt::Display`
   --> library/core/src/fmt/num.rs:195:39
    |
    |
184 | / macro_rules! debug {
185 | |     ($($T:ident)*) => {$(
186 | |         #[stable(feature = "rust1", since = "1.0.0")]
187 | |         impl fmt::Debug for $T {
...   |
195 | |                     fmt::Display::fmt(self, f)
    | |                                       ^^^^ `i64` cannot be formatted with the default formatter
199 | |     )*};
200 | | }
    | |_- in this expansion of `debug!`
    | |_- in this expansion of `debug!`
201 | / debug! {
202 | |   i8 i16 i32 i64 i128 isize
203 | |   u8 u16 u32 u64 u128 usize
    | |_- in this macro invocation
    |
    = help: the trait `fmt::Display` is not implemented for `i64`
    = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
---
              isize
              u128
            and 4 others

error[E0277]: `u64` doesn't implement `fmt::Display`
   --> library/core/src/fmt/num.rs:195:39
184 | / macro_rules! debug {
184 | / macro_rules! debug {
185 | |     ($($T:ident)*) => {$(
186 | |         #[stable(feature = "rust1", since = "1.0.0")]
187 | |         impl fmt::Debug for $T {
...   |
195 | |                     fmt::Display::fmt(self, f)
    | |                                       ^^^^ `u64` cannot be formatted with the default formatter
199 | |     )*};
200 | | }
    | |_- in this expansion of `debug!`
    | |_- in this expansion of `debug!`
201 | / debug! {
202 | |   i8 i16 i32 i64 i128 isize
203 | |   u8 u16 u32 u64 u128 usize
    | |_- in this macro invocation
    |
    = help: the trait `fmt::Display` is not implemented for `u64`
    = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
---
              isize
              u128
            and 4 others

error[E0277]: `u64` doesn't implement `fmt::Display`
     |
     |
1345 |                       write!(f, "{}{}", prefix, integer_part)?;
     |                       |                         |
     |                       |                         `u64` cannot be formatted with the default formatter
     |                       in this macro invocation (#1)
     |
     |
    ::: library/core/src/macros/mod.rs:630:1
     |
630  | / macro_rules! write {
631  | |     ($dst:expr, $($arg:tt)*) => {
     | |                        ------------------------------ in this macro invocation (#2)
633  | |     };
634  | | }
     | |_- in this expansion of `write!` (#1)
     | |_- in this expansion of `write!` (#1)
...
1012 | /     macro_rules! format_args {
1013 | |         ($fmt:expr) => {{ /* compiler built-in */ }};
1014 | |         ($fmt:expr, $($args:tt)*) => {{ /* compiler built-in */ }};
     | |_____- in this expansion of `$crate::format_args!` (#2)
     |
     = help: the trait `fmt::Display` is not implemented for `u64`
     = note: in format strings you may be able to use `{:?}` (or {:#?} for pretty-print) instead
```

</details>