@rust-log-analyzer 2024-07-25
The job **`x86_64-gnu-tools`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/10100788227/job/27932948048) [(plain)](https://github.com/rust-lang/rust/commit/fe7cdd871576e44a317570ea14c90c78e4d0d91b/checks/27932948048/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>