@rust-log-analyzer 2024-07-25
The job **`x86_64-gnu-tools`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/10101254133/job/27934425534) [(plain)](https://github.com/rust-lang/rust/commit/01aa12512eb46c7bf47044d64a12326b6a0b29ed/checks/27934425534/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>