@GuillaumeGomez 2024-07-26
Applied your suggestions (thanks a lot!) and regenerated new performance results. `u64` and `i64` are now equivalent to the previous implementation but the improvements for the other types (`i8` and `u8` in particular) remain. So I suppose it's positive. :)