@rust-log-analyzer 2024-07-26
The job **`x86_64-gnu-tools`** failed! Check out the build log: [(web)](https://github.com/rust-lang/rust/actions/runs/10113568330/job/27970232568) [(plain)](https://github.com/rust-lang/rust/commit/d0144a1a3f457216f4767a296a7c28f109be9beb/checks/27970232568/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>